+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Next version
+* Varanus::SSL::CSR - support EC certs
+
 ### Version 0.5.0 (2021-01-26)
 * Add Varanus::Domain
 * Add Varanus::SSL#list and Varanus::SSL#info